kann benotet werden

+++ src/main/java/Ellipse.java
@@ -1,23 +1,23 @@
 public class Ellipse extends Figure {
-    private double[] coordinates = new double[2];
-    private double mainAxis;
-    private double minorAxis;
+    private float[] coordinates = new float[2];
+    private float mainAxis;
+    private float minorAxis;
 
-    public Ellipse(double[] coordinates, double mainAxis, double minorAxis) {
+    public Ellipse(float[] coordinates, float mainAxis, float minorAxis) {
         this.setCoordinates(coordinates);
         this.setMainAxis(mainAxis);
         this.setMinorAxis(minorAxis);
     }
 
     @Override
-    public double area() {
-        return getMainAxis() * getMinorAxis() * Math.PI;
+    public float area() {
+        return getMainAxis() * getMinorAxis() * (float) Math.PI;
     }
 
     @Override
-    public double circumference() {
-        double lambda = (getMainAxis() - getMinorAxis()) / (getMainAxis() + getMinorAxis());
-        return (getMainAxis() + getMinorAxis()) * Math.PI * (1 + ((3 * lambda * lambda)) / (10 + Math.sqrt(4 - 3 * lambda * lambda)));
+    public float circumference() {
+        float lambda = (getMainAxis() - getMinorAxis()) / (getMainAxis() + getMinorAxis());
+        return (getMainAxis() + getMinorAxis()) * (float) Math.PI * (1 + ((3 * lambda * lambda)) / (10 + (float) Math.sqrt(4 - 3 * lambda * lambda)));
     }
 
     @Override
@@ -26,27 +26,27 @@ public class Ellipse extends Figure {
     }
 
     //region Getter and Setter
-    public double[] getCoordinates() {
+    public float[] getCoordinates() {
         return coordinates;
     }
 
-    public void setCoordinates(double[] coordinates) {
+    public void setCoordinates(float[] coordinates) {
         this.coordinates = coordinates;
     }
 
-    public double getMainAxis() {
+    public float getMainAxis() {
         return mainAxis;
     }
 
-    public void setMainAxis(double mainAxis) {
+    public void setMainAxis(float mainAxis) {
         this.mainAxis = mainAxis;
     }
 
-    public double getMinorAxis() {
+    public float getMinorAxis() {
         return minorAxis;
     }
 
-    public void setMinorAxis(double minorAxis) {
+    public void setMinorAxis(float minorAxis) {
         this.minorAxis = minorAxis;
     }
     //endregion
